+++ kernel_tuner/backends/c.py
@@ -13,7 +13,7 @@ import numpy as np
 import numpy.ctypeslib
 
 from kernel_tuner.backends.backend import CompilerBackend
-from kernel_tuner.observers.observer import BenchmarkObserver
+from kernel_tuner.observers.c import CRuntimeObserver
 from kernel_tuner.util import get_temp_filename, delete_temp_file, write_file, SkippableFailure
 
 dtype_map = {"int8": C.c_int8,
@@ -31,26 +31,6 @@ dtype_map = {"int8": C.c_int8,
 # It contains a numpy object (ndarray or number) and a ctypes object with a copy
 # of the argument data. For an ndarray, the ctypes object is a wrapper for the ndarray's data.
 Argument = namedtuple("Argument", ["numpy", "ctypes"])
-
-
-class CRuntimeObserver(BenchmarkObserver):
-    """ Observer that collects results returned by benchmarking function in the C backend """
-
-    def __init__(self, dev):
-        self.dev = dev
-        self.objective = "time"
-        self.times = []
-
-    def after_finish(self):
-        self.times.append(self.dev.last_result)
-
-    def get_results(self):
-        results = {
-            self.objective: np.average(self.times),
-            self.objective + "s": self.times.copy()
-        }
-        self.times = []
-        return results
 
 
 class CFunctions(CompilerBackend):
