+++ kernel_tuner/backends/opencl.py
@@ -4,29 +4,13 @@ import time
 import numpy as np
 
 from kernel_tuner.backends.backend import GPUBackend
-from kernel_tuner.observers.observer import BenchmarkObserver
+from kernel_tuner.observers.opencl import OpenCLObserver
 
 #embedded in try block to be able to generate documentation
 try:
     import pyopencl as cl
 except ImportError:
     cl = None
-
-
-class OpenCLObserver(BenchmarkObserver):
-    """ Observer that measures time using CUDA events during benchmarking """
-    def __init__(self, dev):
-        self.dev = dev
-        self.times = []
-
-    def after_finish(self):
-        event = self.dev.event
-        self.times.append((event.profile.end - event.profile.start)*1e-6) #ms
-
-    def get_results(self):
-        results = {"time": np.average(self.times), "times": self.times.copy()}
-        self.times = []
-        return results
 
 
 class OpenCLFunctions(GPUBackend):
